Update functional codes
Reflect some minor changes in the administrative organizations for the
City Hall as of October 2019, after the August 2019 changes.

We're trying with the most conservative approach first, by keeping all
the old codes while possible, as we'll move to custom mappings in a case
by case basis upon revision.

civio/presupuesto-management#944